requirements.txt
python-dotenv
requests
ascii-magic
pillow